+++ modules/Book.js
@@ -1,0 +1,7 @@
+export default class Book {
+  constructor(title, author) {
+    this.id = -1;
+    this.title = title;
+    this.author = author;
+  }
+}
