+++ modules/Storage.js
@@ -1,0 +1,9 @@
+const getBooks = () => {
+  return JSON.parse(localStorage.getItem('books')) || [];
+};
+
+const saveBooks = (booksArray) => {
+  localStorage.setItem('books', JSON.stringify(booksArray));
+};
+
+export { getBooks, saveBooks };
